# talentwood-backend-develop/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/home/allmovielist": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get all the movie list",
                "description": "Return All the movie list",
                "operationId": "346f0259aff5c8e95a1d7afab777700b",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/home/movielist": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get movielist which is active for this week",
                "description": "Return All the movie list which is active",
                "operationId": "002f0cc4a2fa55e8d13fd7c6dc3395d5",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/home/banners": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get banner list",
                "description": "Return All the banners list which is active",
                "operationId": "ff495006192ff01f03300a4a72ddf15b",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/home/movielist/{id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get movie based on Id",
                "description": "Return the movie corresponding to the Id",
                "operationId": "451c49a03f264a39c491d1f643115e5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/home/search": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Search movies by name",
                "description": "Search for the movie based on the movie name query string",
                "operationId": "40a39ab80c3f2d9491b55b246ffb7a03",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "search movie name query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/goldenHourMovie": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get The Current Golden Hour Movie",
                "description": "Return the Golden Hour Movie",
                "operationId": "fb912535270eccd90d31461dd74c6900",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    }
}